Move all files out of the linux directory.
They're not Linux-specific.
