--- src/linux/mod.rs
@@ -1,10 +1,0 @@
-// Licensed under the Apache License, Version 2.0, <LICENSE-APACHE or
-// http://apache.org/licenses/LICENSE-2.0> or the MIT license <LICENSE-MIT or
-// http://opensource.org/licenses/MIT>, at your option. This file may not be
-// copied, modified, or distributed except according to those terms.
-
-pub mod elf;
-
-mod lines;
-mod source;
-mod symbol;
