+++ src/symbol.rs
